WIP #17 Cleaned up the code in data_survey_21_merge.Rmd
I've made changes to data_survey_21_merge.Rmd that are necessary to preserve the original SPSS labels where they exist. I still need to:
- Test using the codebook code.
- Add variable labels to created variables that didn't come from SPSS.
- Update the data pulls.

However, these are really outside of the scope of the current issue. I'm going to go ahead and close this issue and do the rest of the work on a separate issue branch.